byte to int

+++ Domain/VBMS.Domain/Enums/Gender.cs
@@ -2,7 +2,7 @@
 
 namespace VBMS.Domain.Enums;
 
-public enum Gender : byte
+public enum Gender : int
 {
     None = 0,
     Male,
